pyproject.toml
[project]
name = "chembl-ai"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "pyarrow>=22.0.0",
    "pyspark[connect]==4.1.0",
    "tinygrad>=0.12.0",
]

[dependency-groups]
dev = [
    "apache-airflow>=3.1.6",
    "ipykernel>=7.1.0",
    "matplotlib>=3.10.8",
    "mlflow>=3.8.1",
]


[tool.ruff]
exclude = ["*.ipynb"]
